+++ src/Client/router/Router.java
@@ -1,9 +1,5 @@
 package client.router;
-
 import java.awt.*;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseMotionAdapter;
 
 import client.GamePanel;
 import client.KeyHandler;
@@ -27,17 +23,15 @@ public class Router {
     private PreviewRoute previewRoute;
     private VictoryRoute victoryRoute;
     private OptionRoute optionRoute;
+    protected GamePanel gamePanel;
 
-    private Rectangle startButtonBounds, optionButtonBounds, exitButtonBounds;
-    private boolean hoverStart, hoverOption, hoverExit;
-    private GamePanel gamePanel;
 
     public Router(GamePanel gamePanel, KeyHandler keyHand) {
         this.gamePanel = gamePanel;
         Router.currentRoute = GamePanel.index;
         this.keyHand = keyHand;
         this.keyHand.setRoute(this);
-        this.lobbyRoute = new LobbyRoute(this);
+        this.lobbyRoute = new LobbyRoute(this, gamePanel);
         this.scenceRoute = new ScenceRoute(this);
         this.characterRoute = new CharacterRouter(this);
         this.itemRoute = new ItemRoute(this);
@@ -45,10 +39,6 @@ public class Router {
         this.victoryRoute = new VictoryRoute(this);
         this.optionRoute = new OptionRoute(this);
 
-        this.startButtonBounds = new Rectangle((gamePanel.screenWidth / 2) - 155, 307 - 70, 290, 100);
-        this.optionButtonBounds = new Rectangle((gamePanel.screenWidth / 2) - 155, 450 - 70, 290, 100);
-        this.exitButtonBounds = new Rectangle((gamePanel.screenWidth / 2) - 155, 593 - 70, 290, 100);
-        setupMouseListener();
     }
 
     public void update() {
@@ -75,49 +65,10 @@ public class Router {
                 optionRoute.update();
                 break;
         }
-    }
-
-    private void setupMouseListener() {
-        gamePanel.addMouseMotionListener(new MouseMotionAdapter() {
-            @Override
-            public void mouseMoved(MouseEvent e) {
-                Point mousePoint = e.getPoint();
-
-                // Check if the mouse is hovering over any button
-                hoverStart = startButtonBounds.contains(mousePoint);
-                hoverOption = optionButtonBounds.contains(mousePoint);
-                hoverExit = exitButtonBounds.contains(mousePoint);
-
-                // Set cursor to hand pointer if hovering over a button
-                if (hoverStart || hoverOption || hoverExit) {
-                    gamePanel.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
-                } else {
-                    gamePanel.setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
-                }
-
-                // Repaint the game panel to reflect hover changes
-                gamePanel.repaint();
-            }
-        });
-
-        gamePanel.addMouseListener(new MouseAdapter() {
-            @Override
-            public void mouseClicked(MouseEvent e) {
-                Point clickPoint = e.getPoint();
-                if (startButtonBounds.contains(clickPoint)) {
-                    System.out.println("Start button clicked!");
-                    Router.currentRoute = Router.S_SCENCE_STATE;
-
-                } else if (optionButtonBounds.contains(clickPoint)) {
-                    System.out.println("Option button clicked!");
-                    Router.currentRoute = Router.OPTION_STATE;
-
-                } else if (exitButtonBounds.contains(clickPoint)) {
-                    System.out.println("Exit button clicked!");
-                    System.exit(0);
-                }
-            }
-        });
-
+        if(currentRoute != LOBBY_STATE){
+            lobbyRoute.startButtonBounds = new Rectangle(0,0);
+            lobbyRoute.optionButtonBounds = new Rectangle(0,0);
+            lobbyRoute.exitButtonBounds= new Rectangle(0,0);
+        }
     }
 }
